+++ bfasst/utils/__init__.py
@@ -1,0 +1,2 @@
+"""Import general functions into the utils namespace"""
+from .general import *
